removes getClassNamesForShabadRatings from utils, and categorizes utils [wip]
