+++ src/js/util/ang/read-ang.ts
@@ -1,0 +1,6 @@
+import {
+  LOCAL_STORAGE_KEY_FOR_PREVIOUSLY_READ_ANG,
+} from '../../constants';
+
+export const readAng = (): number =>
+  parseInt(localStorage.getItem(LOCAL_STORAGE_KEY_FOR_PREVIOUSLY_READ_ANG));
