+++ src/js/util/ang/save-ang.ts
@@ -1,0 +1,8 @@
+import {
+  LOCAL_STORAGE_KEY_FOR_PREVIOUSLY_READ_ANG,
+} from '../../constants';
+
+import { saveToLocalStorage } from '../localstorage';
+
+export const saveAng = (ang: number) =>
+  saveToLocalStorage(LOCAL_STORAGE_KEY_FOR_PREVIOUSLY_READ_ANG, ang.toString());
