+++ src/js/util/ang/should-save-ang.ts
@@ -1,0 +1,11 @@
+interface IShouldSaveAngArguments {
+  type: string,
+  info: {
+    source: {
+      sourceId: string
+    }
+  }
+}
+
+export const shouldSaveAng = ({ type, info }: IShouldSaveAngArguments) =>
+  type === 'ang' && info.source.sourceId === 'G';
