--- src/js/util/get-classnames-for-shabad-ratings.ts
@@ -1,8 +1,0 @@
-export const getClassnamesForShabadRatings = (avgScore: number): string => {
-  if (avgScore >= 70) {
-    return 'Excellent'
-  } else if (avgScore >= 40) {
-    return 'Good';
-  }
-  return 'Poor'
-}
