+++ src/js/util/localstorage/get-string-from-localstorage.ts
@@ -1,0 +1,5 @@
+export const getStringFromLocalStorage = (key: string, defaultValue = null): string => {
+  const value = localStorage.getItem(key);
+  if (value === null) return defaultValue;
+  return value;
+};
