+++ src/js/util/localstorage/index.ts
@@ -1,0 +1,5 @@
+export * from './get-boolean-from-localstorage';
+export * from './get-number-from-localstorage';
+export * from './get-string-from-localstorage';
+export * from './get-array-from-localstorage';
+export * from './save-to-localstorage';
