+++ src/js/util/localstorage/save-to-localstorage.ts
@@ -1,0 +1,4 @@
+//   Saves to localStorage in next animation frame.
+
+export const saveToLocalStorage = (key: string, value: string) =>
+  requestAnimationFrame(() => localStorage.setItem(key, value));
