+++ src/js/util/numbers/index.ts
@@ -1,1 +1,2 @@
 export * from './to-fixed-float';
+export * from './numbers-range';
