+++ src/js/util/shabad/index.ts
@@ -1,0 +1,1 @@
+export * from './get-shabad-list';
