+++ src/js/util/url/get-parameter-by-name.ts
@@ -1,0 +1,8 @@
+export const getParameterByName = (_name: string, url: string = window.location.href): string | undefined => {
+  const name = _name.replace(/[[\]]/g, '\\$&');
+  const regex = new RegExp(`[?&]${name}(=([^&#]*)|&|#|$)`);
+  const results = regex.exec(url);
+  if (!results) return undefined;
+  if (!results[2]) return '';
+  return decodeURIComponent(results[2].replace(/\+/g, ' '));
+}
