+++ src/js/util/url/get-query-params.ts
@@ -1,0 +1,14 @@
+export const getQueryParams = (qString: string = document.location.search): object => {
+  const qs = qString.replace(/\+/g, ' ');
+  const params: any = {};
+  const re = /[?&]?([^=]+)=([^&]*)/g;
+  let tokens;
+
+  while ((tokens = re.exec(qs))) {
+    const firstToken = tokens[1];
+    const secondToken = tokens[2];
+    params[decodeURIComponent(firstToken)] = decodeURIComponent(secondToken);
+  }
+
+  return params;
+}
