+++ src/js/util/url/to-ang-url.ts
@@ -1,0 +1,14 @@
+import { objectToQueryParams } from "..";
+
+interface IToAngURLArguments {
+  ang: number,
+  source: string,
+  highlight: string | number
+}
+
+export const toAngURL = ({ ang, source, highlight }: IToAngURLArguments) =>
+  `/ang?${objectToQueryParams({
+    ang,
+    source,
+    highlight,
+  })}`;
