+++ src/js/util/url/to-shabad-url.ts
@@ -1,0 +1,24 @@
+import { objectToQueryParams } from "..";
+interface IToShabadUrlArguments {
+  shabad: {
+    shabadId: string | number,
+    verseId: string
+  },
+  q: string,
+  type?: string,
+  source?: string
+}
+
+export const toShabadURL = ({
+  shabad: { shabadId: id, verseId: highlight },
+  q,
+  type = undefined,
+  source = undefined,
+}: IToShabadUrlArguments) =>
+  `/shabad?${objectToQueryParams({
+    id,
+    q,
+    type,
+    source,
+    highlight,
+  })}`;
